Added additional doc strings and vim-surround.
There's a new package vim-surround in .emacs. Added a fancy header
from figlet. And more of the file is documented for easy reading and
understanding. In this commit, nord-theme has been removed because I
moved away from the melpa stable version and moved to the master
branch with a custom fork that I use because the stable version
doesn't have syntax highlighting (aka font-locking) enabled for eshell.