Update init.lua to support autocomplete
init.lua is now in a working state. This will be considered neovim
1.0.0 from this point on. There are lots of things to be improved
going forward, mainly involving the lazy-loading of the plugins that
are installed. Things like Goyo stick out as not needing to be loaded
until they're called, since I'm guessing that particular plugin will
only ever be called maybe one or twice a year.